requirements.txt
streamlit
openai
Pillow
pymupdf
python-docx
pandas
pydub
audiorecorder